Fix lstm numpy conversion.

+++ lstm.py
@@ -198,11 +198,11 @@ def try_model(access_key, secret_key, num_layers, l2_reg, dropout, bucket_name, 
             group_sorted = group.sort_values('year')
             for i in range(len(group_sorted) - 1):
                 if group_sorted.iloc[i+1]['year'] - group_sorted.iloc[i]['year'] == 1:
-                    sequence = group_sorted.iloc[i:i+2][input_attrs].values
+                    sequence = group_sorted.iloc[i:i+2][input_attrs].values.astype('float32')
                     X.append(sequence)
-                    y.append(group_sorted.iloc[i+1][output_attrs].values)
+                    y.append(group_sorted.iloc[i+1][output_attrs].values.astype('float32'))
         
-        return numpy.array(X), numpy.array(y)
+        return numpy.array(X, dtype='float32'), numpy.array(y, dtype='float32')
 
     frame = pandas.read_csv(temp_file_path)
     frame['setAssign'] = frame['year'].apply(lambda x:
